Added TELEOP_START in the planner. Added padding for the rotated lever.
TELEOP_START calls GetReadyForTeleop() and goes to initik and then to standik. The operator then can return back home using End Task as usual. This is tested. I also added padding and unpadding for the rotated lever that we asd in the environment for obstacle avoidance. This is tested as well. Left hand turn succeeds and goes back home avoiding collision nicely.